Dump more detail about `Frame` objects when debugging
When logging `Frame` objects, the current code shows the payload as raw
bytes. This means that when debugging the protocol, it is often
necessary to disassemble frames by hand.

This patch provides support for parsing several common packet types and
dumping the parsed versions (at least when parsing succeeds) instead of
the raw bytes.